backend/requirements.txt
# GMTM Agent SDK Backend Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# AI Models
anthropic>=0.40.0
openai>=2.0.0

# Database
pymysql>=1.1.0
mysql-connector-python>=8.2.0

# Web Research
requests>=2.31.0
beautifulsoup4>=4.12.0

# Environment & Utilities
python-dotenv>=1.0.0
httpx>=0.25.0

# CORS
starlette>=0.27.0
cryptography

claude-agent-sdk
